OpenGLSystem and related classes moved to systems/opengl
